pyproject.toml
[project]
name = "signature-streamlit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "langchain-openai>=1.0.2",
    "pillow>=12.0.0",
    "streamlit>=1.51.0",
]
